+++ lib/td_bg/search/aggregations.ex
@@ -33,18 +33,22 @@ defmodule TdBg.Search.Aggregations do
     |> Enum.into(%{})
   end
 
-  def template_terms(%{content: content}) do
+  defp template_terms(%{content: content}) do
     content
     |> Enum.filter(&filter_content_term/1)
-    |> Enum.map(& &1["name"])
+    |> Enum.map(&Map.take(&1, ["name", "type"]))
     |> Enum.map(&content_term/1)
   end
 
-  def filter_content_term(%{"name" => "_confidential"}), do: true
-  def filter_content_term(%{"values" => values}) when is_map(values), do: true
-  def filter_content_term(_), do: false
+  defp filter_content_term(%{"name" => "_confidential"}), do: true
+  defp filter_content_term(%{"values" => values}) when is_map(values), do: true
+  defp filter_content_term(_), do: false
 
-  defp content_term(field) do
+  defp content_term(%{"name" => field, "type" => "user"}) do
+    {field, %{terms: %{field: "content.#{field}.raw", size: 50}}}
+  end
+
+  defp content_term(%{"name" => field}) do
     {field, %{terms: %{field: "content.#{field}.raw"}}}
   end
 end